WIP settings pre-removing references to old settings

+++ Assets/Scripts/Settings/Setting.cs
@@ -2,14 +2,16 @@
 using UnityEngine.UI;
 
 /*
- * represents a single setting, such as a button or slider
+ * Represents a single setting, such as a button or slider.
+ * Assign the font size and size of the setting in the inspector.
+ * Bound the setting size at 160 units to prevent clipping.
  */
 public abstract class Setting : MonoBehaviour {
 
     // Assigned in inspector
     public string id;
     public float settingSize = 130;
-    public float detailsFontSize = 10;
+    public int detailsFontSize = 10;
 
     public abstract void RefreshData();
